Refactor code

+++ src/Math/Combinatorics/combination.jl
@@ -1,5 +1,5 @@
 
 # combinations(a, k) = Base.combinations(a, k)
 
-Base.combinations(a, k, ::Type{Val{:unique}}) = combinations(a, k)
-Base.combinations(a, k, ::Type{Val{:repeated}}) = multicombinations(a, k)
+Base.combinations(a, k::Integer, ::Type{Val{:unique}}) = combinations(a, k)
+Base.combinations(a, k::Integer, ::Type{Val{:repeated}}) = multicombinations(a, k)
